refactor: service for deriv. & integ.
