+++ src/main/java/com/trbaxter/github/fractionalcomputationapi/controller/IndexController.java
@@ -1,5 +1,9 @@
 package com.trbaxter.github.fractionalcomputationapi.controller;
 
+import com.trbaxter.github.fractionalcomputationapi.model.ControllerRequest;
+import com.trbaxter.github.fractionalcomputationapi.service.derivation.CaputoDerivativeService;
+import com.trbaxter.github.fractionalcomputationapi.service.derivation.RiemannLiouvilleDerivativeService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -8,12 +12,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.trbaxter.github.fractionalcomputationapi.model.ControllerRequest;
-import com.trbaxter.github.fractionalcomputationapi.service.CaputoDerivativeService;
-import com.trbaxter.github.fractionalcomputationapi.service.RiemannLiouvilleDerivativeService;
-
-import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("fractional-calculus-computation-api/derivative")
@@ -33,7 +31,7 @@ public class IndexController {
 	@PostMapping("/caputo")
 	public ResponseEntity<String> computeCaputoDerivative(@Valid @RequestBody ControllerRequest request) {
 		try {
-			String result = caputoDerivativeService.computeDerivative(request.getCoefficients(), request.getOrder());
+			String result = caputoDerivativeService.evaluateExpression(request.getCoefficients(), request.getOrder());
 			return new ResponseEntity<>(result, HttpStatus.OK);
 		} catch (Exception e) {
 			return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
@@ -43,7 +41,7 @@ public class IndexController {
 	@PostMapping("/riemann-liouville")
 	public ResponseEntity<String> computeRiemannLiouvilleDerivative(@Valid @RequestBody ControllerRequest request) {
 		try {
-			String result = riemannLiouvilleDerivativeService.computeDerivative(request.getCoefficients(),
+			String result = riemannLiouvilleDerivativeService.evaluateExpression(request.getCoefficients(),
 					request.getOrder());
 			return new ResponseEntity<>(result, HttpStatus.OK);
 		} catch (Exception e) {
