--- src/main/java/com/trbaxter/github/fractionalcomputationapi/service/DerivativeService.java
@@ -1,5 +1,0 @@
-package com.trbaxter.github.fractionalcomputationapi.service;
-
-public interface DerivativeService {
-    String computeDerivative(double[] coefficients, double alpha);
-}
