+++ src/main/java/com/trbaxter/github/fractionalcomputationapi/service/FractionalCalculusService.java
@@ -1,0 +1,5 @@
+package com.trbaxter.github.fractionalcomputationapi.service;
+
+public interface FractionalCalculusService {
+	String evaluateExpression(double[] coefficients, double alpha);
+}
